[skip ci] path utils cleanup
